finalizado

+++ src/controllers/tarefas-controller.ts
@@ -1,7 +1,6 @@
 import { prisma } from "../libs/Prisma"
 import { RequestHandler } from "express"
 import jwt from 'jsonwebtoken'
-import { User } from "../types/user-type"
 
 export const listaTarefa: RequestHandler = async (req, res) => {
     const token = req.headers.authorization?.split(' ')[1]
@@ -19,19 +18,23 @@ export const listaTarefa: RequestHandler = async (req, res) => {
         })
         res.json(tarefas)
     } catch(error){
+        console.log("Authorization Header:", req.headers.authorization);
+        console.log('erro no listar tarefa', error);
         res.status(401).json({msg: "Não autorizado"})
     }
 }
 export const criarTarefa: RequestHandler = async(req, res) => {
-    const token = req.headers.authorization?.split(' ')[1]
-    const decoded = jwt.verify(token as string, process.env.SECRET_JWT as string) as { userId: number }
+    
     
     try{
-        const { title, description } = req.body
+        const token = req.headers.authorization?.split(' ')[1]
+        const decoded = jwt.verify(token as string, process.env.SECRET_JWT as string) as { userId: number }
+        const { title, description, date } = req.body
         const tarefa = await prisma.task.create({
             data: {
                 title,
                 description,
+                date: new Date(date),
                 userId: decoded.userId
             }
         })
@@ -48,25 +51,26 @@ export const criarTarefa: RequestHandler = async(req, res) => {
     }
 
 }
-export const deletarTarefa: RequestHandler = async(req, res) => {
-    try{
-        const { id } = req.params
-        const tarefa = await prisma.task.delete({
-            where: {
-                id: Number(id)
-            }
-        })
-        if(tarefa){
-            res.status(200).json({msg: "tarefa deletada com sucesso"})
-        } else {
-            res.status(400).json({msg: "erro ao deletar tarefa"})
-        }
-    }
-    catch(error){
-        res.status(400).json({msg: "erro ao deletar tarefa"})
-    }
-}
+export const deletarTarefa: RequestHandler = async (req, res, next) => {
+  try {
+    const { id } = req.params;
+    const idNum = Number(id);
 
+    if (isNaN(idNum)) {
+      res.status(400).json({ msg: "ID inválido." });
+      return;
+    }
+
+    await prisma.task.delete({
+      where: { id: idNum },
+    });
+
+    res.status(200).json({ msg: "Tarefa deletada com sucesso." });
+  } catch (error) {
+    console.error("Erro ao deletar tarefa:", error);
+    res.status(400).json({ msg: "Erro ao deletar tarefa." });
+  }
+};
 export const atualizarTarefa: RequestHandler = async (req, res, next) => {
   try {
     const idParam = req.params.id;
